requirements.txt
click
joblib
marshmallow>=3.0.0b12
matplotlib
netCDF4
numpy
requests
torchvision
tqdm
seaborn
scikit-optimize
